refactor: re-export svg assets consider touch area
