+++ src/components/SignUp/StepIcon/StepIcon.tsx
@@ -1,13 +1,13 @@
-"use client"
+"use client";
 
 import { memo } from "react";
 import IcStep1 from "@/assets/icons/ic_step1.svg";
 import IcStep2 from "@/assets/icons/ic_step2.svg";
 import IcStep3 from "@/assets/icons/ic_step3.svg";
 import IcStep4 from "@/assets/icons/ic_step4.svg";
-import IcCheckTransparent from "@/assets/icons/ic_check-transparent.svg";
-import IcGreyCircle from "@/assets/icons/ic_grey-circle.svg";
-import * as styles from './stepIcon.css';
+import IcCheckTransparent from "@/assets/icons/ic_check_transparent.svg";
+import IcGreyCircle from "@/assets/icons/ic_grey_circle.svg";
+import * as styles from "./stepIcon.css";
 
 interface StepIconProps {
   step: 1 | 2 | 3 | 4;
@@ -21,37 +21,27 @@ const STEP_CONFIG = {
 } as const;
 
 export function StepIcon({ step }: StepIconProps) {
-    const renderIcons = () => {
-      return Array(4)
-        .fill(null)
-        .map((_, index) => {
-          const currentStep = index + 1;
+  const renderIcons = () => {
+    return Array(4)
+      .fill(null)
+      .map((_, index) => {
+        const currentStep = index + 1;
 
-          if (currentStep === step) {
-            const StepComponent = STEP_CONFIG[step].Icon;
-            return (
-              <div className={styles.stepIconWrapper} key={index}>
-                <StepComponent />
-                <span className={styles.stepText}>
-                  {STEP_CONFIG[step].text}
-                </span>
-              </div>
-            );
-          }
-
-          return currentStep < step ? (
-            <IcCheckTransparent key={index} />
-          ) : (
-            <IcGreyCircle key={index} />
+        if (currentStep === step) {
+          const StepComponent = STEP_CONFIG[step].Icon;
+          return (
+            <div className={styles.stepIconWrapper} key={index}>
+              <StepComponent />
+              <span className={styles.stepText}>{STEP_CONFIG[step].text}</span>
+            </div>
           );
-        });
-    };
+        }
 
-    return (
-      <div className={styles.stepIconContainer}>
-        {renderIcons()}
-      </div>
-    );
-  }
+        return currentStep < step ? <IcCheckTransparent key={index} /> : <IcGreyCircle key={index} />;
+      });
+  };
 
-  export const MemoizedStepIcon = memo(StepIcon);
+  return <div className={styles.stepIconContainer}>{renderIcons()}</div>;
+}
+
+export const MemoizedStepIcon = memo(StepIcon);
