+++ src/components/SignUp/Terms/AllAgreeCheckboxField/AllAgreeCheckboxField.tsx
@@ -1,6 +1,6 @@
 import React, { memo } from "react";
-import IcCheckComplete from "@/assets/icons/ic_check-complete.svg";
-import IcCheckWhiteGrey from "@/assets/icons/ic_check-whitegrey.svg";
+import IcCheckComplete from "@/assets/icons/ic_check_complete.svg";
+import IcCheckWhiteGrey from "@/assets/icons/ic_check_whitegrey.svg";
 import * as styles from "./allAgreeCheckboxField.css";
 
 interface AllAgreeCheckboxProps {
